[Fix] Error Fix
- 의존한 클래스나 코드가 없는 경우 전부 다 주석 처리
- 일단 실행이 되게 바꿔놓음

+++ src/main/kotlin/org/washcode/washpang/domain/handledItems/service/HandledItemsService.kt
@@ -20,7 +20,7 @@ class HandledItemsService(
 //            }
 //    }
 
-    fun getAllHandledItems(laundryShopId: Long): List<HandledItems> {
+    fun getAllHandledItems(laundryShopId: Int): List<HandledItems> {
         try {
             return handledItemsRepository.findByLaundryshopId(laundryShopId)
         } catch (e: Exception) {
